+++ packages/renderer-worker/src/parts/ViewletMain/ViewletMainCloseOthers.js
