+++ packages/renderer-worker/test/ViewletMain.test.js
@@ -965,1005 +965,6 @@ test.skip('handleTabContextMenu', async () => {
   ])
 })
 
-test('closeOthers - 0 0 - first tab is selected and first tab is focused', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 0, 0)
-})
-
-test('closeOthers - 0 1 - first tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 0, 0)
-})
-
-test('closeOthers - 0 2 - first tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 0, 0)
-})
-
-test('closeOthers - 1 0 - second tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 1, 1)
-})
-
-test('closeOthers - 1 1 - second tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 1, 1)
-})
-
-test('closeOthers - 1 2 - second tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 1, 1)
-})
-
-test('closeOthers - 2 0 - third tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 2, 2)
-})
-
-test('closeOthers - 2 1 - third tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 2, 2)
-})
-
-test('closeOthers - 2 2 - third tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeOthers(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeOthers', 2, 2)
-})
-
-test('closeTabsRight - 0 0 - first tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 0)
-})
-
-test('closeTabsRight - 0 1 - first tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 0)
-})
-
-test('closeTabsRight - 0 2 - first tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 0)
-})
-
-test('closeTabsRight - 1 0 - second tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 1)
-})
-
-test('closeTabsRight - 1 1 - second tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 1)
-})
-
-test('closeTabsRight - 1 2 - second tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 1)
-})
-
-test('closeTabsRight - 2 0 - third tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 2)
-})
-
-test('closeTabsRight - 2 1 - third tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 2)
-})
-
-test('closeTabsRight - 2 2 - third tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsRight(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsRight', 2)
-})
-
-test('closeTabsLeft - 0 0 - first tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 0)
-})
-
-test('closeTabsLeft - 0 1 - first tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 0)
-})
-
-test('closeTabsLeft - 0 2 - first tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 0,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-1.txt',
-    },
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 0)
-})
-
-test('closeTabsLeft - 1 0 - second tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 1)
-})
-
-test('closeTabsLeft - 1 1 - second tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 1)
-})
-
-test('closeTabsLeft - 1 2 - second tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 2,
-    focusedIndex: 1,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-2.txt',
-    },
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 1)
-})
-
-test('closeTabsLeft - 2 0 - third tab is focused and first tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 0,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 2)
-})
-
-test('closeTabsLeft - 2 1 - third tab is focused and second tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 2)
-})
-
-test('closeTabsLeft - 2 2 - third tab is focused and third tab is selected', async () => {
-  // @ts-ignore
-  RendererProcess.invoke.mockImplementation(() => {})
-  const state = {
-    ...ViewletMain.create(1),
-    editors: [
-      {
-        uri: '/test/file-1.txt',
-      },
-      {
-        uri: '/test/file-2.txt',
-      },
-      {
-        uri: '/test/file-3.txt',
-      },
-    ],
-    activeIndex: 1,
-    focusedIndex: 2,
-  }
-  ViewletStates.set('EditorText', {
-    factory: {
-      loadContent() {},
-      refresh() {},
-    },
-    state: {},
-  })
-  await ViewletMain.closeTabsLeft(state)
-  expect(state.editors).toEqual([
-    {
-      uri: '/test/file-3.txt',
-    },
-  ])
-  expect(RendererProcess.invoke).toHaveBeenCalledWith('Viewlet.send', 1, 'closeTabsLeft', 2)
-})
-
 test.skip('resize', () => {
   const state = ViewletMain.create(1)
   const { newState } = ViewletMain.resize(state, {
